Add IMAGE constants
