+++ src/components/list/node.js
@@ -6,8 +6,8 @@ const { Editable } = require('../editable')
 const { Collapse } = require('../fx')
 const { IconFolder, IconGhost, IconTriangle } = require('../icons')
 const { DragSource, DropTarget, NativeTypes, getEmptyImage } = require('../dnd')
-const { DND, LIST, SASS, MIME } = require('../../constants')
-const { isImageSupported } = MIME
+const { DND, LIST, SASS, IMAGE } = require('../../constants')
+const { isImageSupported } = IMAGE
 const { bounds } = require('../../dom')
 const lazy = require('./tree')
 const cx = require('classnames')
