+++ src/components/project/name.js
@@ -4,7 +4,7 @@ const React = require('react')
 const { DropTarget, NativeTypes } = require('../dnd')
 const { IconMaze, IconWarningSm } = require('../icons')
 const { Editable } = require('../editable')
-const { isImageSupported } = require('../../constants/mime')
+const { isImageSupported } = require('../../constants/image')
 const { blank } = require('../../common/util')
 const cx = require('classnames')
 const { bool, func, string } = require('prop-types')
