+++ src/constants/mime.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const MIME = {
+module.exports = {
   GIF: 'image/gif',
   HEIC: 'image/heic',
   HEIC_SEQ: 'image/heic-sequence',
@@ -15,44 +15,3 @@ const MIME = {
   TIFF: 'image/tiff',
   WEBP: 'image/webp'
 }
-
-MIME.EXT = [
-  'gif',
-  'heic',
-  'heif',
-  'jpg',
-  'jpeg',
-  'pdf',
-  'png',
-  'svg',
-  'tif',
-  'tiff',
-  'webp'
-]
-
-const SUPPORTED = {
-  [MIME.GIF]: true,
-  [MIME.HEIC]: true,
-  [MIME.HEIC_SEQ]: true,
-  [MIME.HEIF]: true,
-  [MIME.HEIF_SEQ]: true,
-  [MIME.JPEG]: true,
-  [MIME.PNG]: true,
-  [MIME.PDF]: true,
-  [MIME.SVG]: true,
-  [MIME.TIFF]: true,
-  [MIME.WEBP]: true
-}
-
-MIME.WEB = {
-  [MIME.JPG]: true,
-  [MIME.PNG]: true,
-  [MIME.WEBP]: true,
-  [MIME.GIF]: true,
-  [MIME.SVG]: true
-}
-
-MIME.isImageSupported = (file) =>
-  (typeof file === 'string') ? SUPPORTED[file] : SUPPORTED[file.type]
-
-module.exports = MIME
